+++ Sources/Chaining/UITextView+Chain.swift
@@ -148,7 +148,7 @@ public extension UITextView {
   }
 
   @discardableResult
-  func textColor(_ hex: UInt32, alpha: CGFloat = 1) -> Self {
+  func textColor(_ hex: UInt64, alpha: CGFloat = 1) -> Self {
     textColor(.init(hex: hex))
   }
 }
