+++ Sources/Chaining/UIView+Chain.swift
@@ -638,7 +638,7 @@ public extension UIView {
   }
 
   @discardableResult
-  func backgroundColor(_ hex: UInt32, alpha: CGFloat = 1) -> Self {
+  func backgroundColor(_ hex: UInt64, alpha: CGFloat = 1) -> Self {
     backgroundColor(.init(hex: hex, alpha: alpha))
   }
 }
